refactor: group item extensions
